Reset version to 1.0.0 after fork
This is a fork of Microsoft/elm-json-tree-view,
which I was the author of. In 2019 Microsoft renamed
their Github organization to start with a lowercase 'm'.
This broke Elm's package system as it does not follow
HTTP redirects (and probably would not cope well with
the name change even if it did follow redirects).

So I'm republishing under my personal account.